change emulator structure
 we earlier used to have a single big file with everything.
