+++ emulator/emulator.go
@@ -2,15 +2,12 @@ package main
 
 import (
 	"fmt"
-	"github.com/chzyer/readline"
-	"github.com/fatih/color"
 	"log"
-	"strings"
 
-	consts "github.com/OpenFinancing/openfinancing/consts"
 	database "github.com/OpenFinancing/openfinancing/database"
 	solar "github.com/OpenFinancing/openfinancing/platforms/solar"
 	scan "github.com/OpenFinancing/openfinancing/scan"
+	"github.com/spf13/viper"
 )
 
 // package emulator is used to emulate the environment of the platform and make changes
@@ -18,39 +15,68 @@ import (
 // should be used only for testing.
 
 // have different entities that will be used across the files here
+// emulator is intended to be a model for a frontend platform that would later be developed
+// using the same backend that we have right now
 var (
-	LocalRecipient  database.Recipient
-	LocalUser       database.User
-	LocalInvestor   database.Investor
-	LocalContractor solar.Entity
-	LocalOriginator solar.Entity
-	LocalSeed       string
-	LocalSeedPwd    string
+	LocalRecipient    database.Recipient
+	LocalUser         database.User
+	LocalInvestor     database.Investor
+	LocalContractor   solar.Entity
+	LocalOriginator   solar.Entity
+	LocalSeed         string
+	LocalSeedPwd      string
+	PlatformPublicKey string
 )
 
 var ApiUrl = "http://localhost:8080"
-var PlatformPublicKey = "GDULAIM6N6SIW7MWS3NDJPY3UIFOHSM4766WQ6O6EKFDBC7PF53VKYLY"
 
-func main() {
+// this platform public key would be public, so makes sense that we hardcode it
+
+func SetupConfig() (string, error) {
+	var err error
+	viper.SetConfigType("yaml")
+	viper.SetConfigName("emulator")
+	viper.AddConfigPath(".")
+
+	err = viper.ReadInConfig()
+	if err != nil {
+		log.Println("Error while reading email values from config file")
+		return "", err
+	}
+
+	PlatformPublicKey = viper.Get("PlatformPublicKey").(string)
 
 	fmt.Println("WELCOME TO THE SMARTSOLAR EMULATOR")
+
 	ColorOutput("ENTER YOUR USERNAME: ", CyanColor)
 	username, err := scan.ScanForString()
 	if err != nil {
 		log.Fatal(err)
 	}
+
 	ColorOutput("ENTER YOUR PASSWORD: ", CyanColor)
 	pwhash, err := scan.ScanForPassword()
 	if err != nil {
 		log.Fatal(err)
 	}
+
 	// need to validate with the RPC here
-	wString, err := Login(username, pwhash)
+	role, err := Login(username, pwhash)
+	if err != nil {
+		return "", err
+	}
+	return role, nil
+}
+
+func main() {
+
+	role, err := SetupConfig()
 	if err != nil {
 		log.Fatal(err)
 	}
-
-	switch wString {
+	switch role {
+	// start loops for each role, would be nice if we could come up with an alternative to
+	// duplication here
 	case "Investor":
 		log.Fatal(LoopInv())
 	case "Recipient":
@@ -62,178 +88,4 @@ func main() {
 	default:
 		log.Println("It should never come here")
 	}
-}
-
-func LoopInv() error {
-	// this loop is for an investor
-	// we have authenticated the user and stored the details in an appropriate structure
-	// need to repeat this struct everywhere because having separate functions and importing
-	// it doesn't seem to work
-	// the problem with having a conditional statement inside the loop is that it checks
-	// role each time and that's not nice performance wise
-	// TOOD: look at alternatives if possible
-	promptColor := color.New(color.FgHiYellow).SprintFunc()
-	whiteColor := color.New(color.FgHiWhite).SprintFunc()
-	rl, err := readline.NewEx(&readline.Config{
-		Prompt:      promptColor("emulator") + whiteColor("# "),
-		HistoryFile: consts.TellerHomeDir + "/history.txt",
-		// AutoComplete: lc.NewAutoCompleter(),
-	})
-
-	ColorOutput("YOUR SEED IS: "+LocalSeed, RedColor)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer rl.Close()
-
-	for {
-		// setup reader with max 4K input chars
-		msg, err := rl.Readline()
-		if err != nil {
-			log.Println(err)
-			return err
-		}
-		msg = strings.TrimSpace(msg)
-		if len(msg) == 0 {
-			continue
-		}
-		rl.SaveHistory(msg)
-
-		cmdslice := strings.Fields(msg)
-		ColorOutput("entered command: "+msg, YellowColor)
-
-		err = ParseInputInv(cmdslice)
-		if err != nil {
-			log.Println(err)
-			return err
-		}
-	}
-	return nil
-}
-
-func LoopRecp() error {
-	// This loop is exclusive to a recipient
-	promptColor := color.New(color.FgHiYellow).SprintFunc()
-	whiteColor := color.New(color.FgHiWhite).SprintFunc()
-	rl, err := readline.NewEx(&readline.Config{
-		Prompt:      promptColor("emulator") + whiteColor("# "),
-		HistoryFile: consts.TellerHomeDir + "/history.txt",
-		// AutoComplete: lc.NewAutoCompleter(),
-	})
-
-	ColorOutput("YOUR SEED IS: "+LocalSeed, RedColor)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer rl.Close()
-
-	for {
-		// setup reader with max 4K input chars
-		msg, err := rl.Readline()
-		if err != nil {
-			log.Println(err)
-			return err
-		}
-		msg = strings.TrimSpace(msg)
-		if len(msg) == 0 {
-			continue
-		}
-		rl.SaveHistory(msg)
-
-		cmdslice := strings.Fields(msg)
-		ColorOutput("entered command: "+msg, YellowColor)
-
-		err = ParseInputRecp(cmdslice)
-		if err != nil {
-			log.Println(err)
-			return err
-		}
-	}
-	return nil
-}
-
-func LoopOrig() error {
-	// This loop is exclusive to an originator
-	promptColor := color.New(color.FgHiYellow).SprintFunc()
-	whiteColor := color.New(color.FgHiWhite).SprintFunc()
-	rl, err := readline.NewEx(&readline.Config{
-		Prompt:      promptColor("emulator") + whiteColor("# "),
-		HistoryFile: consts.TellerHomeDir + "/history.txt",
-		// AutoComplete: lc.NewAutoCompleter(),
-	})
-
-	ColorOutput("YOUR SEED IS: "+LocalSeed, RedColor)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer rl.Close()
-
-	for {
-		// setup reader with max 4K input chars
-		msg, err := rl.Readline()
-		if err != nil {
-			log.Println(err)
-			return err
-		}
-		msg = strings.TrimSpace(msg)
-		if len(msg) == 0 {
-			continue
-		}
-		rl.SaveHistory(msg)
-
-		cmdslice := strings.Fields(msg)
-		ColorOutput("entered command: "+msg, YellowColor)
-
-		err = ParseInputOrig(cmdslice)
-		if err != nil {
-			log.Println(err)
-			return err
-		}
-	}
-	return nil
-}
-
-func LoopCont() error {
-	// This loop is exclusive to a contractor
-	promptColor := color.New(color.FgHiYellow).SprintFunc()
-	whiteColor := color.New(color.FgHiWhite).SprintFunc()
-	rl, err := readline.NewEx(&readline.Config{
-		Prompt:      promptColor("emulator") + whiteColor("# "),
-		HistoryFile: consts.TellerHomeDir + "/history.txt",
-		// AutoComplete: lc.NewAutoCompleter(),
-	})
-
-	ColorOutput("YOUR SEED IS: "+LocalSeed, RedColor)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer rl.Close()
-
-	for {
-		// setup reader with max 4K input chars
-		msg, err := rl.Readline()
-		if err != nil {
-			log.Println(err)
-			return err
-		}
-		msg = strings.TrimSpace(msg)
-		if len(msg) == 0 {
-			continue
-		}
-		rl.SaveHistory(msg)
-
-		cmdslice := strings.Fields(msg)
-		ColorOutput("entered command: "+msg, YellowColor)
-
-		err = ParseInputCont(cmdslice)
-		if err != nil {
-			log.Println(err)
-			return err
-		}
-	}
-	return nil
 }
