+++ emulator/emulator.yaml
@@ -1,0 +1,2 @@
+# this file can be in git tracking since it does not contain any secret
+PlatformPublicKey: "GDULAIM6N6SIW7MWS3NDJPY3UIFOHSM4766WQ6O6EKFDBC7PF53VKYLY"
